+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Breaking changes
-- End of support for 1.21.6-1.21.8. Nine Lifes now requires 1.21.9+
+- End of support for 1.21.6-1.21.9. Nine Lifes now requires 1.21.10+
+- Rewrite config to YACL instead of MidnightLib
 
 ## Optimisations
 - Use Fabric Api events for bed sleeping problems instead of Mixin's
+
+## Dev
+- Bump fabric loom version to 1.15.2